add useMembersColumns hook
