+++ src/components/ui/members-table.tsx
@@ -1,7 +1,6 @@
 import {
   ColumnFiltersState,
   VisibilityState,
-  createColumnHelper,
   flexRender,
   getCoreRowModel,
   getFilteredRowModel,
@@ -19,10 +18,8 @@ import {
 
 import { useEffect, useMemo, useState } from "react";
 import { useTranslation } from "react-i18next";
-import { getCustomDate, hasExpired } from "@/lib/utils";
 import { Skeleton } from "./skeleton";
 import InfiniteScroll from "react-infinite-scroll-component";
-import { StatusBadge } from "./status-badge";
 import { SearchBar } from "./searchbar";
 import { Separator } from "@radix-ui/react-separator";
 import { useMembersMutations } from "@/hooks/use-table-mutations";
@@ -31,13 +28,7 @@ import { useWindowSize } from "@/hooks/use-window-size";
 import { useMembersQuery } from "@/hooks/use-members-query";
 import { FilterPopover } from "./filter-popover";
 import { HideFieldsPopover } from "./hide-fields-popover";
-import { ActionButtons } from "./action-buttons";
-
-const columnHelper = createColumnHelper<Member>();
-const membersPerPage = 20;
-interface Row {
-  original: Member;
-}
+import { useMembersColumns } from "@/hooks/use-members-columns";
 
 export type Member = {
   id: number;
@@ -61,10 +52,11 @@ export type Member = {
   note: string;
 };
 
+const membersPerPage = 20;
+
 export function DataTable() {
   const { t } = useTranslation();
-  const { renewMutation, updateMutation, insertMutation } =
-    useMembersMutations();
+  const { insertMutation } = useMembersMutations();
   const isMobile = useWindowSize();
   const [debouncedSearch, setDebouncedSearch] = useState<string | null>(null);
   const [columnFilters, setColumnFilters] = useState<ColumnFiltersState>([]);
@@ -97,83 +89,7 @@ export function DataTable() {
   }, [refetch, debouncedSearch]);
 
   // Build table
-  const columns = useMemo(
-    () => [
-      columnHelper.accessor((row) => `${row.name} ${row.surname}`, {
-        id: "fullName",
-        meta: t("membersTable.name"),
-        cell: (info) => info.getValue(),
-        header: () => <span>{t("membersTable.name")}</span>,
-      }),
-      columnHelper.accessor("email", {
-        meta: t("membersTable.email"),
-        cell: (info) => info.getValue(),
-        header: () => <span>{t("membersTable.email")}</span>,
-      }),
-      columnHelper.accessor("birthDate", {
-        meta: t("membersTable.birthDate"),
-        cell: (info) => getCustomDate(info.getValue()),
-        header: () => <span>{t("membersTable.birthDate")}</span>,
-      }),
-      columnHelper.accessor("cardNumber", {
-        meta: t("membersTable.cardNumber"),
-        cell: (info) => {
-          const result = info.getValue();
-          return result ? result : "-";
-        },
-        header: () => <span>{t("membersTable.cardNumber")}</span>,
-      }),
-      columnHelper.accessor("expirationDate", {
-        meta: t("membersTable.expirationDate"),
-        cell: (info) => {
-          const result = getCustomDate(info.getValue());
-          return result ? result : "-";
-        },
-        header: () => <span>{t("membersTable.expirationDate")}</span>,
-        filterFn: (row, columnId) => {
-          return hasExpired(new Date(row.getValue(columnId)));
-        },
-      }),
-      columnHelper.accessor("status", {
-        meta: t("membersTable.status"),
-        cell: (info) => <StatusBadge status={info.getValue()} />,
-        header: () => <span>{t("membersTable.status")}</span>,
-        filterFn: "equals",
-      }),
-      columnHelper.accessor("suspendedTill", {
-        meta: t("membersTable.suspendedTill"),
-        cell: (info) => {
-          const result = getCustomDate(info.getValue());
-          return result ? result : "-";
-        },
-        header: () => <span>{t("membersTable.suspendedTill")}</span>,
-        filterFn: (row, columnId) => {
-          const cellValue = row.getValue(columnId);
-          return cellValue ? !hasExpired(new Date(cellValue as string)) : false;
-        },
-      }),
-      columnHelper.accessor("isActive", {
-        meta: t("membersTable.isActive"),
-        header: () => <span>{t("membersTable.isActive")}</span>,
-        filterFn: "equals",
-      }),
-      columnHelper.accessor("isDeleted", {
-        meta: t("membersTable.isDeleted"),
-        header: () => <span>{t("membersTable.isDeleted")}</span>,
-        filterFn: "equals",
-      }),
-      {
-        meta: t("membersTable.actions"),
-        id: "actions",
-        header: () => <span className="ml-3">{t("membersTable.actions")}</span>,
-        cell: ({ row }: { row: Row }) => {
-          return <ActionButtons row={row.original} />;
-        },
-      },
-    ],
-    [t, renewMutation, updateMutation],
-  );
-
+  const columns = useMembersColumns();
   const members = useMemo(() => {
     return data?.pages.reduce<Member[]>((acc, page) => {
       return [...acc, ...page.members];
